requirements.txt
python-dotenv
langchain
langchain-openai
langchain-core
langchain-google-genai
langchain-anthropic
langchain_google_firestore